use size with proper type (for 64bit) for size_t and pointer components

+++ trunk/source/CACSL2BoogieTranslator/src/de/uni_freiburg/informatik/ultimate/cdt/translation/implementation/base/expressiontranslation/BitvectorTranslation.java
@@ -624,11 +624,7 @@ public class BitvectorTranslation extends ExpressionTranslation {
 
 	@Override
 	public CPrimitive getCTypeOfPointerComponents() {
-		// 2015-10-29 Matthias: using int is unsound on 64bit systems, but it
-		// probably saves a lot of conversions and I guess this unsoundness
-		// is never a problem in the SV-COMP and most other code
-		// 2020-12-08 Dominik: But this does lead to type errors, for instance in the aws benchmarks in SV-COMP.
-		return new CPrimitive(CPrimitives.UINT);
+		return new CPrimitive(CPrimitives.ULONG);
 	}
 
 	@Override
